requirements.txt
attrs==23.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
cssselect==1.2.0
cssutils==2.11.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
importlib_metadata==8.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
keyring==25.2.1
lxml==5.2.2
more-itertools==10.3.0
outcome==1.3.0.post0
packaging==24.1
premailer==3.10.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
SecretStorage==3.3.3
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tomli==2.0.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yagmail==0.15.293
zipp==3.19.2
